Изменения: 1. Убран закомментированный код 2. Метод создания HTTP запроса для URLRequest вынесено расширением в отдельный файл 3. Небольшие правки
