+++ ImageFeed/Services/OAuth2/OAuth2Service.swift
@@ -70,52 +70,3 @@ extension OAuth2Service {
         )
     }
 }
-
-extension URLRequest {
-    static func makeHTTPRequest(
-        path: String,
-        httpMethod: String,
-        baseUrl: URL = defaultBaseUrl
-    ) -> URLRequest {
-        var request = URLRequest(url: URL(string: path, relativeTo: baseUrl)!)
-        request.httpMethod = httpMethod
-        return request
-    }
-}
-
-//private enum NetworkError: Error {
-//    case httpStatusCode(Int)
-//    case urlRequestError(Error)
-//    case urlSessionError
-//}
-//
-//extension URLSession {
-//    func data(
-//        for request: URLRequest,
-//        completion: @escaping (Result<Data, Error>) -> Void
-//    ) -> URLSessionDataTask {
-//        let fulfillCompletion: (Result<Data, Error>) -> Void = { result in
-//            DispatchQueue.main.async {
-//                completion(result)
-//            }
-//        }
-//        let task = dataTask(with: request, completionHandler: { data, response, error in
-//            if let data = data,
-//               let response = response,
-//               let statusCode = (response as? HTTPURLResponse)?.statusCode
-//            {
-//                if 200..<300 ~= statusCode {
-//                    fulfillCompletion(.success(data))
-//                } else {
-//                    fulfillCompletion(.failure(NetworkError.httpStatusCode(statusCode)))
-//                }
-//            } else if let error = error {
-//                fulfillCompletion(.failure(NetworkError.urlRequestError(error)))
-//            } else {
-//                fulfillCompletion(.failure(NetworkError.urlSessionError))
-//            }
-//        })
-//        task.resume()
-//        return task
-//    }
-//}
